refactor: 주문 엔티티 상태 enum을 enums 패키지로 별도 분리
