+++ src/main/java/com/project/baedalsodae/order/entity/enums/ActorType.java
@@ -1,4 +1,4 @@
-package com.project.baedalsodae.order.entity;
+package com.project.baedalsodae.order.entity.enums;
 
 public enum ActorType {
 	SYSTEM, CUSTOMER, OWNER, MANAGER, MASTER
